requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
click==8.1.6
comm==0.1.4
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
Flask==2.2.5
graphviz
gunicorn==21.2.0
imageio==2.31.1
ipykernel==6.16.2
ipython
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jupyter_core
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow
platformdirs==3.10.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
six==1.16.0
stack-data==0.6.2
tornado
traitlets==5.9.0
wcwidth==0.2.6
Werkzeug